Rename fastify.types.ts to env.types.ts
Remove the last Fastify naming vestige. The file only contains
BoilrEnv (the environment type for context variables and bindings),
so env.types.ts is a more accurate name. Updates all 15 imports.
