+++ packages/core/src/core/file-routes/file-routes.ts
@@ -2,7 +2,7 @@ import path from "node:path";
 import type { OpenAPIHono } from "@hono/zod-openapi";
 import type { AuthConfig } from "../../types/auth.types.js";
 import type { ExceptionConfig } from "../../types/error.types.js";
-import type { BoilrEnv } from "../../types/fastify.types.js";
+import type { BoilrEnv } from "../../types/env.types.js";
 import type { FileRoutesOptions } from "../../types/file-routes.types.js";
 import { registerRoutes } from "./route-loader.js";
 import { extractRouteInfo, scanDirectories } from "./scanner.js";
