+++ packages/core/src/core/server.ts
@@ -1,6 +1,6 @@
 import type { OpenAPIHono } from "@hono/zod-openapi";
 import type { MiddlewareHandler } from "hono";
-import type { BoilrEnv } from "../types/fastify.types.js";
+import type { BoilrEnv } from "../types/env.types.js";
 import type { BoilrConfig } from "./config.js";
 
 export interface BoilrInstance {
