+++ packages/core/src/plugins/monitoring.plugin.ts
@@ -1,6 +1,6 @@
 import type { MiddlewareHandler } from "hono";
 import type { BoilrConfig, BoilrMonitorConfig } from "../core/config.js";
-import type { BoilrEnv } from "../types/fastify.types.js";
+import type { BoilrEnv } from "../types/env.types.js";
 
 export const createMonitorMiddleware = (config: BoilrConfig): MiddlewareHandler<BoilrEnv> => {
   const defaultOptions: BoilrMonitorConfig = {
